requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Text processing and embeddings
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0
datasets>=2.0.0
accelerate>=0.20.0

# BibTeX parsing
bibtexparser>=1.4.0
chardet>=3.0.4

# Clustering algorithms
umap-learn>=0.5.3
hdbscan>=0.8.29
python-igraph>=0.10.0
leidenalg>=0.9.0
python-louvain>=0.16

# Network analysis
networkx>=2.8.0

# Topic modeling
bertopic>=0.15.0

# Web requests and progress
requests>=2.28.0
tqdm>=4.64.0

# Jupyter notebooks
jupyter>=1.0.0
jupyterlab>=3.4.0
notebook>=6.4.0

# Template rendering
jinja2>=3.1.0

# Utilities
pathlib2>=2.3.0
